+++ detectors/node/opentelemetry-resource-detector-azure/src/detectors/index.ts
@@ -15,3 +15,5 @@
  */
 
 export * from './AzureAppServiceDetector';
+export * from './AzureFunctionsDetector';
+export * from './AzureVmDetector';