+++ plugins/node/instrumentation-tedious/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.34.3 to ^0.35.0
+
 ## [0.6.3](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-tedious-v0.6.2...instrumentation-tedious-v0.6.3) (2023-11-13)