+++ plugins/node/instrumentation-tedious/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.34.3 to ^0.35.0
+
+## [0.6.5](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-tedious-v0.6.4...instrumentation-tedious-v0.6.5) (2024-01-04)
+
+
+### Bug Fixes
+
+* **deps:** update otel core experimental ([#1866](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1866)) ([9366543](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/9366543f5572e1e976ce176ddeb0b438f6c16c45))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.35.0 to ^0.35.1
+
 ## [0.6.3](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-tedious-v0.6.2...instrumentation-tedious-v0.6.3) (2023-11-13)